Hone down examples to index.js files
